refactor: extract vaadin-dialog-overlay to separate file
